app/src/test/java/com/example/chatsunset/SettingsActivityTest.kt

import android.os.Looper
import com.example.chatsunset.activities.SettingsActivity
import com.example.chatsunset.models.User
import com.google.android.gms.tasks.Task
import com.google.android.gms.tasks.Tasks
import com.google.firebase.FirebaseApp
import com.google.firebase.auth.FirebaseAuth
import com.google.firebase.auth.FirebaseUser
import com.google.firebase.firestore.CollectionReference
import com.google.firebase.firestore.DocumentReference
import com.google.firebase.firestore.DocumentSnapshot
import com.google.firebase.firestore.FirebaseFirestore
import com.google.firebase.storage.FirebaseStorage
import com.google.firebase.storage.StorageReference
import junit.framework.TestCase.assertEquals
import junit.framework.TestCase.assertNotNull
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import org.mockito.ArgumentMatchers.anyMap
import org.mockito.Mockito.mock
import org.mockito.Mockito.`when`
import org.robolectric.Robolectric
import org.robolectric.RobolectricTestRunner
import org.robolectric.RuntimeEnvironment
import org.robolectric.Shadows
import org.robolectric.annotation.Config
import org.robolectric.shadows.ShadowLooper

@RunWith(RobolectricTestRunner::class)
@Config(sdk = [28])
class SettingsActivityTest {

    private lateinit var activity: SettingsActivity
    private lateinit var auth: FirebaseAuth
    private lateinit var firestore: FirebaseFirestore
    private lateinit var storage: FirebaseStorage
    private lateinit var userDocumentReference: DocumentReference
    private lateinit var updateTask: Task<Void>
    @Before
    fun setUp() {
        FirebaseApp.initializeApp(RuntimeEnvironment.getApplication())
        auth = mock(FirebaseAuth::class.java)
        firestore = mock(FirebaseFirestore::class.java)
        storage = mock(FirebaseStorage::class.java)
        userDocumentReference = mock(DocumentReference::class.java)
        updateTask = mock(Task::class.java) as Task<Void>
        val mockUser = mock(FirebaseUser::class.java)
        `when`(mockUser.uid).thenReturn("123")
        `when`(auth.currentUser).thenReturn(mockUser)
        val collectionReference = mock(CollectionReference::class.java)
        `when`(firestore.collection("users")).thenReturn(collectionReference)
        `when`(collectionReference.document("123")).thenReturn(userDocumentReference)
        `when`(storage.reference).thenReturn(mock(StorageReference::class.java))
        `when`(userDocumentReference.update(anyMap())).thenReturn(updateTask)
        activity = Robolectric.buildActivity(SettingsActivity::class.java).create().get()
        activity.auth = auth
        activity.db = firestore
        activity.storage = storage
        Shadows.shadowOf(Looper.getMainLooper()).idle()
        activity.setupDefaultImagePicker()
    }

    @Test
    fun `activity should be successfully created`() {
        assertNotNull(activity)
    }

    @Test
    fun `interest checkboxes should be populated correctly`() {
        // Call the populate method directly only if not already called during activity setup
        activity.interestCheckboxes.removeAllViews() // Clear previous views to ensure clean state
        activity.populateInterestsCheckboxes()
        assertEquals(6, activity.interestCheckboxes.childCount)
    }
//    @Test
//    fun `user data should be updated correctly`() {
//        // Create a user with the expected data
//        val user = User("123", "email@example.com", "username", null, listOf("Sport", "Music"))
//
//        // Ensure the Firestore mock returns the correct document reference
//        `when`(firestore.collection("users").document(user.uuid)).thenReturn(userDocumentReference)
//        `when`(userDocumentReference.update(anyMap())).thenReturn(updateTask)
//
//        // Simulate the success of the update task
//        `when`(updateTask.addOnSuccessListener(any<OnSuccessListener<Void>>())).thenAnswer { invocation ->
//            val onSuccessListener = invocation.arguments[0] as OnSuccessListener<Void>
//            onSuccessListener.onSuccess(null)
//            updateTask
//        }
//
//        // Debugging: Print user values before saving
//        println("Saving user data: ${user.pseudo}, ${user.image}, ${user.interests}")
//
//        // Save user data
//        activity.saveUserData(user)
//
//        // Verify the updateUserData method is called with correct arguments
//        verify(userDocumentReference).update(mapOf(
//            "pseudo" to "username",
//            "image" to "",
//            "interests" to listOf("Sport", "Music")
//        ))
//    }
    @Test
    fun `user data should be loaded correctly`() {
        // Mock FirebaseUser and DocumentReference
        val mockUser = mock(FirebaseUser::class.java)
        val docRef = mock(DocumentReference::class.java)
        val snapshot = mock(DocumentSnapshot::class.java)

        // Set up expected behavior for mocks
        `when`(mockUser.uid).thenReturn("123")
        `when`(auth.currentUser).thenReturn(mockUser)
        `when`(firestore.collection("users").document("123")).thenReturn(docRef)
        `when`(docRef.get()).thenReturn(Tasks.forResult(snapshot))
        `when`(snapshot.toObject(User::class.java)).thenReturn(User("123", "email@example.com", "username", null, listOf("Sport", "Music")))

        // Temporarily enable the email field for testing purposes
        activity.layoutTextInputEmail.isEnabled = true

        // Call loadUserData and process any pending tasks
        activity.loadUserData()
        ShadowLooper.runUiThreadTasksIncludingDelayedTasks()

        // Verify the data is loaded into the activity
        assertEquals("username", "username")

        // Disable the email field again to match the original state
        activity.layoutTextInputEmail.isEnabled = false
    }
    @Test
    fun `firebase should be initialized correctly`() {
        activity.initializeFirebase()
        assertNotNull(activity.auth)
        assertNotNull(activity.db)
        assertNotNull(activity.storage)
    }
}
